requirements.txt
python-dotenv>=1.0.0
tiktoken>=0.5.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.0
dashscope>=1.14.0
chromadb>=0.4.0
rank-bm25>=0.2.2
jieba>=0.42.0
tqdm>=4.65.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
tabulate>=0.9.0